requirements.txt
numpy
pandas
matplotlib
tensorflow
ipykernel
requests
tqdm
pillow
streamlit